Create readme
Prediction of the drug targets of diabetes-related is very important for the pathogenesis of diseases and the development of new drugs. In order to improve the identification accuracy of drug targets in network pharmacology, a novel diabetes-related target prediction algorithm based on target-related compound identification is proposed in this paper. In our proposed method, data mining method is utilized to collect disease-related targets verified by biology experiments from the up-to-date literatures. For each target, the relevant active compounds are collected and a set of inactive compounds are generated by the proposed negative compounds selected algorithm in order to constitute the training set. Molecular descriptors are utilized to extract the feature of each compound, which are input into logistic regression (LR) algorithm to learn and predict compounds of new target. The prediction accuracy is utilized to judge diabetes-related targets. The experiments results show that LR method could identify diabetes-related targets more accurately than other eight classical classifiers.